requirements.txt
pymupdf
anthropic
reportlab
pillow
